backend/requirements.txt
django==5.0.8
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2
psycopg2-binary==2.9.9
Pillow==10.3.0
django-filter==24.2
django-cors-headers==4.3.1
django-storages==1.14.4
boto3==1.35.43
six==1.16.0
requests==2.32.3